+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [UIBULKED-267](https://issues.folio.org/browse/UIBULKED-267) User's names in "Run by" column of Bulk edit logs.
 * [UIBULKED-277](https://issues.folio.org/browse/UIBULKED-277) Birth date column shows time.
 * [UIBULKED-223](https://issues.folio.org/browse/UIBULKED-223) [FE] [TS] Add types for all backend entities
+* [UIBULKED-244](https://issues.folio.org/browse/UIBULKED-244) Add Suppress from discovery option
 
 ## [3.0.5](https://github.com/folio-org/ui-bulk-edit/tree/v3.0.5) (2023-03-22)